scripts/requirements.txt
pandas==2.1.4
python-docx==1.0.1
openpyxl==3.1.2  # For Excel support
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6  # For file upload support 